split: Upload.tsx를 index.tsx와 style.ts로 분리
css를 style.ts로 분리
