+++ src/pages/Upload/style.ts
@@ -1,0 +1,67 @@
+import styled from '@emotion/styled';
+
+export const NextIcon = styled.img`
+  width: auto;
+  height: auto;
+`;
+
+export const HelpSub = styled.span`
+  align-items: center;
+  font-size: 16px;
+  font-weight: bold;
+`;
+
+export const HelpBox = styled.div`
+  border: 1px solid #d2d2dc;
+  width: 100%;
+  height: 90px;
+  background-color: #f5f5f5;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  gap: 10px;
+  padding-left: 30px;
+  border-radius: 5px;
+`;
+
+export const HelpTitle = styled.div`
+  font-size: 24px;
+  font-weight: bold;
+  margin-bottom: 24px;
+`;
+
+export const HelpLinks = styled.span`
+  display: flex;
+  flex-direction: column;
+  margin-top: 65px;
+  margin-bottom: 65px;
+
+  width: 684px;
+`;
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: auto;
+  align-items: center;
+  margin-top: 56px;
+  margin-right: 250px;
+`;
+
+export const PageHeader = styled.div`
+  margin-bottom: 30px;
+  text-align: center;
+`;
+
+export const Title = styled.span`
+  text-align: center;
+  font-weight: bold;
+  font-size: 40px;
+  color: #000000;
+`;
+
+export const Subtitle = styled.p`
+  color: #8f8f8f;
+  font-size: 22px;
+  margin-top: 32px;
+`;
